+++ models/ConfigBuilder.py
@@ -9,6 +9,7 @@ class ConfigBuilder():
     def __init__(self) -> None:
         self._b = 0
         self._c = 0
+        self._k = 0
         self._t = 0
 
     def init(self) -> None:
@@ -129,7 +130,65 @@ class ConfigBuilder():
             else:
                 config['binance']['config']['live'] = 0
 
-        if self._b == 1 or self._c == 1:
+        choice = input("Do you want to use the Kucoin exchange (1=yes, 2=no:default)? ")
+        if choice == '1':
+            self._K = 1
+            config['kucoin'] = {}
+            config['kucoin']['api_url'] = 'https://api.kucoin.com'
+
+            choice = input("Do you have API keys for the Kucoin exchange (1=yes, 2=no:default)? ")
+            if choice == '1':
+                while 'api_key' not in config['kucoin']:
+                    api_key = input("What is your Kucoin API Key? ")
+                    p = re_compile(r"^[a-f0-9]{32,32}$")
+                    if p.match(api_key):
+                        config['kucoin']['api_key'] = api_key
+
+                while 'api_secret' not in config['kucoin']:
+                    api_secret = input("What is your Kucoin API Secret? ")
+                    p = re_compile(r"^[A-z0-9+\/]+==$")
+                    if p.match(api_secret):
+                        config['kucoin']['api_secret'] = api_secret
+
+                while 'api_passphrase' not in config['kucoin']:
+                    api_passphrase = input("What is your Kucoin API Passphrase? ")
+                    p = re_compile(r"^[a-z0-9]{10,11}$")
+                    if p.match(api_passphrase):
+                        config['kucoin']['api_passphrase'] = api_passphrase
+            else:
+                config['kucoin']['api_key'] = '<fill in>'
+                config['kucoin']['api_secret'] = '<fill in>'
+                config['kucoin']['api_passphrase'] = '<fill in>'
+
+            config['kucoin']['config'] = {}
+
+            while 'base_currency' not in config['kucoin']['config']:
+                base_currency = input("What is your Kucoin base currency (what you are buying) E.g. BTC? ")
+                p = re_compile(r"^[A-Z0-9]{3,7}$")
+                if p.match(base_currency):
+                    config['kucoin']['config']['base_currency'] = base_currency
+
+            while 'quote_currency' not in config['kucoin']['config']:
+                quote_currency = input("What is your Kucoin quote currency (what you are buying with) E.g. GBP? ")
+                p = re_compile(r"^[A-Z0-9]{3,7}$")
+                if p.match(quote_currency):
+                    config['kucoin']['config']['quote_currency'] = quote_currency
+
+            choice = input("Do you want to smart switch between 1 hour and 15 minute intervals (1=yes:default, 2=no)? ")
+            if choice == '2':
+                while 'granularity' not in config['kucoin']['config']:
+                    choice = input("What granularity do you want to trade? (60, 300, 900, 3600, 21600, 86400)? ")
+                    if int(choice) in [60, 300, 900, 3600, 21600, 86400]:
+                        config['kucoin']['config']['granularity'] = int(choice)
+
+
+            choice = input("Do you want to start live trading? (1=live, 2=test:default)? ")
+            if choice == '1':
+                config['kucoin']['config']['live'] = 1
+            else:
+                config['kucoin']['config']['live'] = 0
+
+        if self._b == 1 or self._c == 1 or self._k == 1:
             choice = input("Do you have a Telegram Token and Client ID (1=yes, 2=no:default)? ")
             if choice == '1':
                 self._t = 1
@@ -153,6 +212,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['sellatloss'] = 1
                 if self._b == 1:
                     config['binance']['config']['sellatloss'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['sellatloss'] = 1
 
             choice = input("Do you want to sell at the next resistance? (1:yes:default, 2=no)? ")
             if choice != '2':
@@ -160,6 +221,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['sellatresistance'] = 1
                 if self._b == 1:
                     config['binance']['config']['sellatresistance'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['sellatresistance'] = 1
 
             choice = input("Do you only want to trade in a bull market SMA50 > SMA200? (1:yes, 2=no:default)? ")
             if choice != '1':
@@ -167,6 +230,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['disablebullonly'] = 1
                 if self._b == 1:
                     config['binance']['config']['disablebullonly'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['disablebullonly'] = 1
 
             choice = input("Do you want to avoid buying when the price is too high? (1:yes:default, 2=no)? ")
             if choice != '2':
@@ -174,6 +239,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['disablebuynearhigh'] = 1
                 if self._b == 1:
                     config['binance']['config']['disablebuynearhigh'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['disablebuynearhigh'] = 1
 
             choice = input("Do you want to disable the On-Balance Volume (OBV) technical indicator on buys? (1:yes:default, 2=no)? ")
             if choice != '2':
@@ -181,6 +248,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['disablebuyobv'] = 1
                 if self._b == 1:
                     config['binance']['config']['disablebuyobv'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['disablebuyobv'] = 1
 
             choice = input("Do you want to disable the Elder-Ray Index on buys? (1:yes:default, 2=no)? ")
             if choice != '2':
@@ -188,6 +257,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['disablebuyelderray'] = 1
                 if self._b == 1:
                     config['binance']['config']['disablebuyelderray'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['disablebuyelderray'] = 1
 
             choice = input("Do you want to disable saving the CSV tracker on buy and sell events? (1:yes:default, 2=no)? ")
             if choice != '2':
@@ -195,6 +266,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['disabletracker'] = 1
                 if self._b == 1:
                     config['binance']['config']['disabletracker'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['disabletracker'] = 1
 
             choice = input("Do you want to disable writing to the log file? (1:yes, 2=no:default)? ")
             if choice != '2':
@@ -202,6 +275,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['disablelog'] = 1
                 if self._b == 1:
                     config['binance']['config']['disablelog'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['disablelog'] = 1
 
             choice = input("Do you want the bot to auto restart itself on failure? (1:yes:default, 2=no)? ")
             if choice != '2':
@@ -209,6 +284,8 @@ class ConfigBuilder():
                     config['coinbasepro']['config']['autorestart'] = 1
                 if self._b == 1:
                     config['binance']['config']['autorestart'] = 1
+                if self._k == 1:
+                    config['kucoin']['config']['autorestart'] = 1
 
             try:
                 with open('./config.json', 'w') as fout :